+++ ServiceControlManager/TuningSourcesModel.h
@@ -21,7 +21,7 @@ public:
 
 private:
 	std::vector<Network::TuningSourceInfoState> m_sources;
-	std::vector<int> m_sourcesErrorCount;
+	std::vector<qint64> m_sourcesErrorCount;
 
 	inline static const Columns m_columns =
 	{
